move schema error types to schema
