+++ packages/safe-fn/src/safe-fn.test-d.ts
@@ -3,11 +3,11 @@ import { assert, describe, expectTypeOf, test } from "vitest";
 import { z } from "zod";
 import { type ActionResult } from "./result";
 import { createSafeFn } from "./safe-fn-builder";
+import type { TSafeFnDefaultCatchHandlerErr } from "./types/error";
 import type {
-  TSafeFnDefaultCatchHandlerErr,
   TSafeFnInputParseError,
   TSafeFnOutputParseError,
-} from "./types/error";
+} from "./types/schema";
 import type { TPrettify } from "./types/util";
 
 const schemaPrimitive = z.string();
