+++ packages/safe-fn/src/types/error.ts
@@ -1,6 +1,4 @@
 import type { Err, Result } from "neverthrow";
-import type { z } from "zod";
-import type { TSafeFnInput, TSafeFnOutput } from "./schema";
 
 /*
 ################################
@@ -36,48 +34,3 @@ export type TSafeFnDefaultCatchHandlerErr = Err<
     cause: "An uncaught error occurred. You can implement a custom error handler by using `catch()`";
   }
 >;
-
-/**
- * @param TSchema a Zod schema or undefined
- * @param TAsAction indicates weather the error will be returned in an error.
- * These types need to be differentiated by `TAsAction` as `Error` classes can not be sent over the wire in server actions.
- */
-export type TSafeFnParseError<
-  TSchema extends z.ZodTypeAny,
-  TAsAction extends boolean,
-> = TAsAction extends true
-  ? {
-      formattedError: z.ZodFormattedError<z.input<TSchema>>;
-      flattenedError: z.typeToFlattenedError<z.input<TSchema>>;
-    }
-  : z.ZodError<z.input<TSchema>>;
-// Temp
-export type TSafeFnParseErrorNoZod<
-  TSchemaInput,
-  TAsAction extends boolean,
-> = TAsAction extends true
-  ? {
-      formattedError: z.ZodFormattedError<TSchemaInput>;
-      flattenedError: z.typeToFlattenedError<TSchemaInput>;
-    }
-  : z.ZodError<TSchemaInput>;
-
-export type TSafeFnInputParseError<
-  TInputSchema extends TSafeFnInput,
-  TAsAction extends boolean,
-> = TInputSchema extends z.ZodTypeAny
-  ? {
-      code: "INPUT_PARSING";
-      cause: TSafeFnParseError<TInputSchema, TAsAction>;
-    }
-  : never;
-
-export type TSafeFnOutputParseError<
-  TOutputSchema extends TSafeFnOutput,
-  TAsAction extends boolean,
-> = TOutputSchema extends z.ZodTypeAny
-  ? {
-      code: "OUTPUT_PARSING";
-      cause: TSafeFnParseError<TOutputSchema, TAsAction>;
-    }
-  : never;
