+++ packages/safe-fn/src/types/schema.ts
@@ -125,3 +125,48 @@ export type TSchemaOutputOrFallback<
   : TSchema extends z.ZodTypeAny
     ? z.output<TSchema>
     : TFallback;
+
+/**
+ * @param TSchema a Zod schema or undefined
+ * @param TAsAction indicates weather the error will be returned in an error.
+ * These types need to be differentiated by `TAsAction` as `Error` classes can not be sent over the wire in server actions.
+ */
+export type TSafeFnParseError<
+  TSchema extends z.ZodTypeAny,
+  TAsAction extends boolean,
+> = TAsAction extends true
+  ? {
+      formattedError: z.ZodFormattedError<z.input<TSchema>>;
+      flattenedError: z.typeToFlattenedError<z.input<TSchema>>;
+    }
+  : z.ZodError<z.input<TSchema>>;
+// Temp
+export type TSafeFnParseErrorNoZod<
+  TSchemaInput,
+  TAsAction extends boolean,
+> = TAsAction extends true
+  ? {
+      formattedError: z.ZodFormattedError<TSchemaInput>;
+      flattenedError: z.typeToFlattenedError<TSchemaInput>;
+    }
+  : z.ZodError<TSchemaInput>;
+
+export type TSafeFnInputParseError<
+  TInputSchema extends TSafeFnInput,
+  TAsAction extends boolean,
+> = TInputSchema extends z.ZodTypeAny
+  ? {
+      code: "INPUT_PARSING";
+      cause: TSafeFnParseError<TInputSchema, TAsAction>;
+    }
+  : never;
+
+export type TSafeFnOutputParseError<
+  TOutputSchema extends TSafeFnOutput,
+  TAsAction extends boolean,
+> = TOutputSchema extends z.ZodTypeAny
+  ? {
+      code: "OUTPUT_PARSING";
+      cause: TSafeFnParseError<TOutputSchema, TAsAction>;
+    }
+  : never;
